pruning: improve prior assignation
In the previous implementation,
the priors are added before the rotation
taken into account the possible state after rotation.

Now they added after the rotation,
taken into account the state before the rotation.
The effects of the change are minimum,
but they provide a better modeling of the prior,
because they follow the time line
more closely.