infer/diffusion: align the prior assignation with the diffusion process
In the previous implementation
the prior for ancestral points was applied to source pixels.
Now,
it is better coupled with the theoretical expectation
of the probability of the trip
(i.e. the diffusion)
times the prior of the destination pixel.

The effects are relatively minor in computation cost
(a bit more of memory is used to precalculate the priors),
but also now takes into account the prior
of arriving to a given terminal.

Also the code for the prior assignation
is more transparent.