infer/diffusion: refactor rotation of logLike fields
In the previous algorithm,
if after rotation a pixel has multiple destinations,
the likelihood is divided equally between destinations.
Also,
if a destination is already occupied,
then both likelihoods are sum.

While this procedure "conserve" the sum of the likelihoods,
it produce highly discontinuous probability fields.

In the new implementation,
the continuity is preferred,
so if a pixel has multiple destinations,
all destinations will receive the same likelihood
(i.e. not divided among destination pixels).
Also,
if the destination pixel is already occupied,
the maximum value is the one assigned.

With this new procedure,
the likelihood field is more smooth.