infer/diffusion: not scale likelihood
In the previous implementation the likelihood
of a pixel was scaled by the total amount
of pixels in the destination set
as is done in Quintero et al. (2015).

Now,
it use the raw sum of the pixels
without the scaling,
that seems to be more in line
with the developments of Bouckaert (2016),
and the main pruning algorithm
of Felsenstein (1981).