phygeo: change timepix to landscape
In the older version,
paleolandscape models are labeled as time pixelations,
which is the name used in earth package.
But in earth package a time pixelation might include
files with continuous variation,
not only landscape models.

Then the change is made to reduce any possible confusion
on the kind of model expected by PhyGeo.