infer/diffusion: use a distance matrix to speed up the likelihood calculation
It also use an array for the storing of conditional likelihoods.
The changes speed up the calculations by about 5x,
without increasing the complexity of the algorithms.

The cost,
is an increase in the memory consumption,
as a distance matrix for an e360 pixelation
will take approximately 1.4gb of memory.