cmd/phygeo: change richness calculation for 'diff map' command
In the previous implementation
each particle in a time stage are added,
and then the KDE is performed.
But this operation does not reflect the richness
at the time stage,
as in some cases,
a lineage with a restricted posterior,
will produce high richness in that restricted part
of the space,
even if there is only a single lineage.

The new implementation,
kept lineages separated,
and the uses the KDE for this lineages
(just in an default usage of map command),
and then sums the estimation of each lineage
in each pixel,
producing a relative indication
of the richness
in each time slice.