infer/diffusion: take into account diffusion with very slow concentration parameters
To speed up the analysis,
the exponential of an scaled the log likelihood
is precalculated,
and multiplied by the PDF of the spherical normal.
But in the case of a large concentration parameter
(either because a concentration is large,
a branch is small,
or a combination of both),
and the source and destination pixels are quite separated,
then the PDF of the spherical normal are too small,
so when making the multiplication
the value is smaller than the minimum non zero value,
and then a zero is assigned.
After calculating all pixel values,
the probability field of the node could be restricted
and if there is no overlap with the probability field
of its sister node,
then a zero field is assigned to the parent node,
and all of its ancestors,
therefore any calculation is impossible.

To prevent this behavior,
if the sum part of the likelihood calculation
for a pixel is zero,
then the maximum value
(calculated as the sum of the log likelihood of the destination pixel
and the log probability of the spherical normal)
is used.
This value underestimates the likelihood
(as it does not make the whole sum),
but if there are other pixels with a sum,
this values will be far larger than the likelihood of this pixel
(that was so small that was previously set to zero).
and it is unlikely that they will be part of any reconstruction.
On the other hand,
it the whole probability field is made with maximum values
it is expected that it will approximate the overall likelihood
values
(Bouckaert 2016).