cmd/phygeo: use xmap as replacement of mapcmd
The old map command had many responsibilities:
- it draws maps of the stochastic maps
- it calculates the KDE from the stochastic maps

This means that any mapping procedure was slow,
as many calculations must be done,
each time that the command was run,
in particular for large data sets,
and when using KDE procedures.

The new architecture,
separates the calculation of the posteriors
(including KDE)
on the command freq.
The advantage,
is that it reduces the number of calculations
and enhance the reusability of this calculations
(as there are text files that can be read
by other applications).

And now,
the new responsibility of map command
is to draw maps.