infer/diffusion: condition diffusion on an arrival pixels
This is a major change in the structure
of the inference process.

Previously,
the assignment was made using the full spherical normal,
but as many pixels cannot be assigned
(prior is zero),
the probability does not add to 1,
so it is biased towards central parts of areas
with valid pixels.
This also means that the diffusion rates
are usually underestimated.

Now,
the probability is conditioned
by the availability of valid pixels,
which means that particles in some pixels
will be more likely to move relative to other pixels,
even if the distances are equal.
This is the same as done in predefined area methods.
This is also most consistent
with the particle simulation,
in which particles that fall into unavailable pixels
are discarded.

This also means that the probability should be scaled
for each pixel,
and then the calculations will be slightly slower,
although most of the time is really used
for the loop over all possible pixels that was done before,
so hopefully the extra cost is not too steep.