cmd/phygeo: use a single file for distribution range data
In the previous implementation,
the distribution range files
were divided in present-absence pixels
and continuous range maps.
While the most probable situation
is using only a single type of distributional data,
it might be possible that there are circumstances
in which one might to combine the two kind of data,
for example,
when using points,
but one of the terminals is poorly sampled,
or their records are doubtful,
it might be preferable to use an approximate range map
for that terminal
and using points for the remaining terminals.