infer/walk: remove scaling of walk attempts
When all walkers fail to reach any valid pixel,
the program attempts another try,
and do that until there is some success.
In doing this,
the original implementation divide the pixel likelihood
by the number of attempts
(to take into account the total number of walker particles).
While more "correct",
this approach has fewer numerical convergence
as in difficult cases
(i.e. a sample far away from any valid point),
many attempts would be tried,
so the overall likelihood was biased by the number of attempts.

Now,
it just make another try,
but ignore the number of attempts.
It is expected to be more convergent,
as the repeat runs will be more or less in the same values
(values,
that we already know are quite small).