infer/walk: reframe the walk model to be more similar to the diffusion command
Changes are made on the architecture of the random walk
so it will be easier to add the traits Q matrix.

It also set the number of steps to a constant,
instead of changing with the relaxed parameter.
So the relaxed parameter now changes the amount of probability
that left a pixel at each step.
This will make the walk model to be more similar
in the parameter definition to the diffusion model.

The new definition was only implemented
for the down-pass,
so all up-pass functions
were deactivated.